Use latest NCodec API changes.

+++ dse/modelc/model/ncodec.c
@@ -70,7 +70,7 @@ static size_t stream_write(NCODEC* nc, uint8_t* data, size_t len)
     return len;
 }
 
-static int stream_seek(NCODEC* nc, size_t pos, int op)
+static long stream_seek(NCODEC* nc, size_t pos, int op)
 {
     NCodecInstance* _nc = (NCodecInstance*)nc;
     if (_nc && _nc->stream) {
@@ -105,7 +105,7 @@ static int stream_seek(NCODEC* nc, size_t pos, int op)
     return -ENOSTR;
 }
 
-static size_t stream_tell(NCODEC* nc)
+static long stream_tell(NCODEC* nc)
 {
     NCodecInstance* _nc = (NCodecInstance*)nc;
     if (_nc && _nc->stream) {
